MVC Furniture Factory - Working CRUD for Products, Users, Orders Placing Order with sessionStorage, Searching Products WebApi More functionality TODO: Autocomplete etc.

+++ MVC5MovieStore/MVC5MovieStore/Models/ShoppingCart.cs
@@ -137,7 +137,7 @@ namespace MVC5MovieStore.Models
                     Quantity = item.Count
                 };
                 // Set the order total of the shopping cart
-                orderTotal += (item.Count * item.Movie.Price);
+                orderTotal += (double)(item.Count * item.Movie.Price);
  
                 storeDB.OrderDetails.Add(orderDetail);
